+++ smithy-protocols-traits/src/main/resources/META-INF/smithy/smithy.protocols.rpcv2.smithy
@@ -2,17 +2,32 @@ $version: "2.0"
 
 namespace smithy.protocols
 
+use smithy.api#httpError
+use smithy.api#cors
+use smithy.api#endpoint
+use smithy.api#hostLabel
+
+/// An RPC protocol with support for multiple wire formats.
 @trait(selector: "service")
-@protocolDefinition(traits: ["smithy.api#httpError"])
-@documentation("An RPC-based protocol with binary encoding support.")
+@protocolDefinition(traits: [
+    httpError
+    cors
+    endpoint
+    hostLabel
+])
 structure smithyRpcV2 {
-    http: StringList,
-    eventStreamHttp: StringList,
+    /// Priority ordered list of supported HTTP protocol versions.
+    http: StringList
+    /// Priority ordered list of supported HTTP protocol versions
+    /// that are required when using event streams.
+    eventStreamHttp: StringList
+
+    /// Priority ordered list of supported wire formats.
     @required
     format: StringList
 }
 
-@documentation("A list of String shapes.")
+/// A list of String shapes.
 list StringList {
     member: String
 }
